+CHG updated Pixel, Signal, Image classes and simple_digitization model

+++ pyxel/physics/pixel.py
@@ -23,5 +23,5 @@ class Pixel(Array):
         """
         super().__init__()
         self.exp_type = np.int
-        self.type_list = [np.int, np.int16, np.int32, np.int64]        # uint16, uint64 ??
+        self.type_list = [np.int32, np.int64, np.uint32, np.uint64]
         self._array = np.zeros((geo.row, geo.col), dtype=self.exp_type)
